requirements.txt
# Core dependencies
setuptools==68.2.2
wheel==0.42.0

# Document processing
python-docx==0.8.11
PyPDF2==3.0.0

# Web interface
streamlit==1.28.0
click==8.1.7

# Data processing
pandas==2.0.3
numpy==1.24.3

# NLP and ML
nltk==3.8.1
scikit-learn==1.3.0

# Environment and testing
python-dotenv==1.0.0
pytest==7.4.0

# Code quality
black==23.7.0
flake8==6.1.0

# AI/ML
transformers==4.31.0
torch==2.0.1
sentencepiece==0.1.99
protobuf==3.20.3

# Image processing
Pillow==10.0.0 